refactor: Move interface implementation close to struct definition
