+++ api/v1alpha2/generic_rrset.go
@@ -13,8 +13,6 @@
 package v1alpha2
 
 import (
-	"fmt"
-	"strings"
 	"time"
 
 	"k8s.io/apimachinery/pkg/api/meta"
@@ -49,118 +47,6 @@ type GenericRRset interface {
 	SetZoneNotAvailable(zoneName string)
 	SetSynchronizationFailed(lastUpdateTime *metav1.Time, err error)
 	SetAvailable(lastUpdateTime *metav1.Time, name string)
-}
-
-// +kubebuilder:object:root:false
-// +kubebuilder:object:generate:false
-var _ GenericRRset = &RRset{}
-
-func (c *RRset) GetObjectMeta() *metav1.ObjectMeta {
-	return &c.ObjectMeta
-}
-
-func (c *RRset) GetKind() string {
-	return "RRset"
-}
-
-func (c *RRset) GetTypeMeta() *metav1.TypeMeta {
-	return &c.TypeMeta
-}
-
-func (c *RRset) GetSpec() *RRsetSpec {
-	return &c.Spec
-}
-
-func (c *RRset) GetStatus() RRsetStatus {
-	return c.Status
-}
-
-func (c *RRset) SetStatus(status RRsetStatus) {
-	c.Status = status
-}
-
-func (c *RRset) Copy() GenericRRset {
-	return c.DeepCopy()
-}
-
-func (c *RRset) GetDomain() string {
-	return fmt.Sprintf("%s.", strings.TrimSuffix(c.Spec.ZoneRef.Name, "."))
-}
-
-func (c *RRset) SetMissingZone(err error) {
-	setMissingZone(&c.Status, c.Generation, err)
-}
-
-func (c *RRset) SetZoneNotAvailable(zoneName string) {
-	setZoneNotAvailable(&c.Status, c.Generation, zoneName)
-}
-
-func (c *RRset) SetDuplicated(lastUpdateTime *metav1.Time, name string) {
-	setRRsetDuplicated(&c.Status, c.Generation, lastUpdateTime, name)
-}
-
-func (c *RRset) SetSynchronizationFailed(lastUpdateTime *metav1.Time, err error) {
-	setRRsetSynchronizationFailed(&c.Status, c.Generation, lastUpdateTime, err)
-}
-
-func (c *RRset) SetAvailable(lastUpdateTime *metav1.Time, name string) {
-	setRRsetAvailable(&c.Status, c.Generation, lastUpdateTime, name)
-}
-
-// +kubebuilder:object:root:false
-// +kubebuilder:object:generate:false
-var _ GenericRRset = &ClusterRRset{}
-
-func (c *ClusterRRset) GetObjectMeta() *metav1.ObjectMeta {
-	return &c.ObjectMeta
-}
-
-func (c *ClusterRRset) GetKind() string {
-	return "ClusterRRset"
-}
-
-func (c *ClusterRRset) GetTypeMeta() *metav1.TypeMeta {
-	return &c.TypeMeta
-}
-
-func (c *ClusterRRset) GetSpec() *RRsetSpec {
-	return &c.Spec
-}
-
-func (c *ClusterRRset) GetStatus() RRsetStatus {
-	return c.Status
-}
-
-func (c *ClusterRRset) SetStatus(status RRsetStatus) {
-	c.Status = status
-}
-
-func (c *ClusterRRset) Copy() GenericRRset {
-	return c.DeepCopy()
-}
-
-func (c *ClusterRRset) GetDomain() string {
-	return fmt.Sprintf("%s.", strings.TrimSuffix(c.Spec.ZoneRef.Name, "."))
-}
-
-func (c *ClusterRRset) SetMissingZone(err error) {
-	setMissingZone(&c.Status, c.Generation, err)
-}
-
-func (c *ClusterRRset) SetZoneNotAvailable(zoneName string) {
-	setZoneNotAvailable(&c.Status, c.Generation, zoneName)
-}
-
-func (c *ClusterRRset) SetDuplicated(lastUpdateTime *metav1.Time, name string) {
-	setRRsetDuplicated(&c.Status, c.Generation, lastUpdateTime, name)
-}
-
-func (c *ClusterRRset) SetSynchronizationFailed(lastUpdateTime *metav1.Time, err error) {
-	setRRsetSynchronizationFailed(&c.Status, c.Generation, lastUpdateTime, err)
-}
-
-func (c *ClusterRRset) SetAvailable(lastUpdateTime *metav1.Time, name string) {
-	setRRsetAvailable(&c.Status, c.Generation, lastUpdateTime, name)
 }
 
 func setMissingZone(status *RRsetStatus, generation int64, err error) {
