+++ api/v1alpha2/generic_zone.go
@@ -48,94 +48,6 @@ type GenericZone interface {
 	SetAvailable(zoneRes *powerdns.Zone)
 }
 
-// +kubebuilder:object:root:false
-// +kubebuilder:object:generate:false
-var _ GenericZone = &Zone{}
-
-func (c *Zone) GetObjectMeta() *metav1.ObjectMeta {
-	return &c.ObjectMeta
-}
-
-func (c *Zone) GetKind() string {
-	return "Zone"
-}
-
-func (c *Zone) GetTypeMeta() *metav1.TypeMeta {
-	return &c.TypeMeta
-}
-
-func (c *Zone) GetSpec() *ZoneSpec {
-	return &c.Spec
-}
-
-func (c *Zone) GetStatus() ZoneStatus {
-	return c.Status
-}
-
-func (c *Zone) SetStatus(status ZoneStatus) {
-	c.Status = status
-}
-
-func (c *Zone) Copy() GenericZone {
-	return c.DeepCopy()
-}
-
-func (c *Zone) SetDuplicated() {
-	setZoneDuplicated(&c.Status, c.Generation)
-}
-
-func (c *Zone) SetSynchronizationFailed(err error) {
-	setZoneSynchronizationFailed(&c.Status, c.Generation, err)
-}
-
-func (c *Zone) SetAvailable(zoneRes *powerdns.Zone) {
-	setZoneAvailable(&c.Status, c.Generation, zoneRes)
-}
-
-// +kubebuilder:object:root:false
-// +kubebuilder:object:generate:false
-var _ GenericZone = &ClusterZone{}
-
-func (c *ClusterZone) GetObjectMeta() *metav1.ObjectMeta {
-	return &c.ObjectMeta
-}
-
-func (c *ClusterZone) GetKind() string {
-	return "ClusterZone"
-}
-
-func (c *ClusterZone) GetTypeMeta() *metav1.TypeMeta {
-	return &c.TypeMeta
-}
-
-func (c *ClusterZone) GetSpec() *ZoneSpec {
-	return &c.Spec
-}
-
-func (c *ClusterZone) GetStatus() ZoneStatus {
-	return c.Status
-}
-
-func (c *ClusterZone) SetStatus(status ZoneStatus) {
-	c.Status = status
-}
-
-func (c *ClusterZone) Copy() GenericZone {
-	return c.DeepCopy()
-}
-
-func (c *ClusterZone) SetDuplicated() {
-	setZoneDuplicated(&c.Status, c.Generation)
-}
-
-func (c *ClusterZone) SetSynchronizationFailed(err error) {
-	setZoneSynchronizationFailed(&c.Status, c.Generation, err)
-}
-
-func (c *ClusterZone) SetAvailable(zoneRes *powerdns.Zone) {
-	setZoneAvailable(&c.Status, c.Generation, zoneRes)
-}
-
 func setZoneDuplicated(status *ZoneStatus, generation int64) {
 	status.SyncStatus = ptr.To(FAILED_STATUS)
 	status.ObservedGeneration = &generation
